Ok, it works ... but has some missing logic (he's dumb for now)
Seems that App works correctly but there is a problem connected to the logic implementation.
Not all cases are covered by logic part (ClassicChessLogic.groovy) and honestly
I do not want to think about specific logic, that part is some kind of business-logic for me.
I've created all the tools to implement that specific logic, if one day one would like to
finish it - you're welcome :D But for now, I'm happy with the results.